+++ app/src/main/java/com/techcourse/AppWebApplicationInitializer.java
@@ -3,8 +3,8 @@ package com.techcourse;
 import jakarta.servlet.ServletContext;
 import jakarta.servlet.ServletRegistration;
 import nextstep.mvc.DispatcherServlet;
-import nextstep.mvc.controller.tobe.AnnotationHandlerMapping;
-import nextstep.mvc.controller.tobe.ExecuteHandlerAdapter;
+import nextstep.mvc.handler.AnnotationHandlerMapping;
+import nextstep.mvc.adapter.ExecuteHandlerAdapter;
 import nextstep.web.WebApplicationInitializer;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
