+++ app/src/main/java/com/techcourse/controller/LogInOutController.java
@@ -47,6 +47,6 @@ public class LogInOutController {
     public ModelAndView logout(HttpServletRequest request, HttpServletResponse response) {
         final HttpSession session = request.getSession();
         session.removeAttribute(UserSession.SESSION_KEY);
-        return new ModelAndView(new JspView("redirect:/"));
+        return new ModelAndView(new JspView(Pages.INDEX.redirectPageName()));
     }
 }
