+++ app/src/main/java/com/techcourse/repository/InMemoryUserRepository.java
@@ -5,7 +5,7 @@ import java.util.Map;
 import java.util.Optional;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.atomic.AtomicLong;
-import nextstep.mvc.controller.exception.BadRequestException;
+import nextstep.mvc.exception.BadRequestException;
 import nextstep.web.annotation.Repository;
 
 @Repository
