+++ mvc/src/main/java/nextstep/mvc/DispatcherServlet.java
@@ -4,9 +4,11 @@ import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import nextstep.mvc.controller.exception.AbstractCustomException;
-import nextstep.mvc.controller.tobe.HandlerAdapterRegistry;
-import nextstep.mvc.controller.tobe.HandlerMappingRegistry;
+import nextstep.mvc.adapter.HandlerAdapter;
+import nextstep.mvc.exception.AbstractCustomException;
+import nextstep.mvc.handler.HandlerMapping;
+import nextstep.mvc.register.HandlerAdapterRegistry;
+import nextstep.mvc.register.HandlerMappingRegistry;
 import nextstep.mvc.view.JspView;
 import nextstep.mvc.view.ModelAndView;
 import nextstep.mvc.view.View;
