+++ mvc/src/main/java/nextstep/mvc/adapter/ExecuteHandlerAdapter.java
@@ -1,8 +1,8 @@
-package nextstep.mvc.controller.tobe;
+package nextstep.mvc.adapter;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import nextstep.mvc.HandlerAdapter;
+import nextstep.mvc.handler.HandlerExecution;
 import nextstep.mvc.view.ModelAndView;
 
 public class ExecuteHandlerAdapter implements HandlerAdapter {
