+++ mvc/src/main/java/nextstep/mvc/exception/AbstractCustomException.java
@@ -1,4 +1,4 @@
-package nextstep.mvc.controller.exception;
+package nextstep.mvc.exception;
 
 public abstract class AbstractCustomException extends RuntimeException implements CustomException {
