+++ mvc/src/main/java/nextstep/mvc/handler/HandlerKey.java
@@ -1,4 +1,4 @@
-package nextstep.mvc.controller.tobe;
+package nextstep.mvc.handler;
 
 import jakarta.servlet.http.HttpServletRequest;
 import java.util.Objects;
