+++ mvc/src/test/java/nextstep/mvc/scanner/AnnotationHandlerMappingTest.java
@@ -1,7 +1,9 @@
-package nextstep.mvc.controller.tobe;
+package nextstep.mvc.scanner;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
+import nextstep.mvc.handler.AnnotationHandlerMapping;
+import nextstep.mvc.handler.HandlerExecution;
 import nextstep.mvc.view.ModelAndView;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
